requirements.txt
selenium==4.24.0
requests>=2.31.0,<3

pyautogui
pywebview>=4.4
google-genai